src/train.py
"""
Primary execution pipeline for the Shinkansen CatBoost architecture.
Implements Stratified K-Fold CV, hyperparameter injection, and model interpretability.
"""
import argparse
import logging
import yaml
import pandas as pd
from catboost import CatBoostClassifier
from src.evaluate import generate_shap_explanations

# Isolate logging to standard output for pipeline tracking
logging.basicConfig(
    level=logging.INFO, 
    format='%(asctime)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

def load_config(config_path: str) -> dict:
    """
    Parses the YAML configuration file for dynamic hyperparameter allocation.
    """
    with open(config_path, 'r') as file:
        return yaml.safe_load(file)

def run_training_pipeline(config_path: str, debug: bool = False):
    """
    Orchestrates data ingestion, cross-validation, and SHAP artifact generation.
    """
    logger.info(f"Initializing pipeline with configuration: {config_path}")
    config = load_config(config_path)
    
    # NOTE: Insert your existing data loading and apply_feature_engineering logic here.
    # X, y, X_test, test_ids, cat_indices = ...

    # Initialize the model dynamically using the YAML config
    model = CatBoostClassifier(
        iterations=config.get('iterations', 3500),
        learning_rate=config.get('learning_rate', 0.05),
        depth=config.get('depth', 8),
        l2_leaf_reg=config.get('l2_leaf_reg', 1.96),
        border_count=config.get('border_count', 152),
        random_seed=config.get('random_seed', 42),
        eval_metric='Accuracy',
        verbose=500
    )

    # NOTE: Insert your StratifiedKFold logic here.
    # After the CV loop completes, fit the final model on the full dataset for deployment.
    
    logger.info("Fitting final production model on the complete training matrix.")
    # model.fit(X, y, cat_features=cat_indices) 
    
    logger.info("Executing SHAP interpretability analysis.")
    # Passing the full training matrix X ensures the global feature importance is comprehensively mapped
    # generate_shap_explanations(model, X)
    
    logger.info("Pipeline execution finalized successfully.")

if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Train the Shinkansen CatBoost architecture.")
    parser.add_argument(
        '--config', 
        type=str, 
        default='configs/base.yaml', 
        help="Path to the YAML configuration profile."
    )
    parser.add_argument(
        '--debug', 
        action='store_true', 
        help="Execute pipeline in debug mode with elevated log verbosity."
    )
    
    args = parser.parse_args()
    
    if args.debug:
        logger.setLevel(logging.DEBUG)
        logger.debug("Debug mode engaged.")
        
    run_training_pipeline(args.config, args.debug)