requirements.txt
catboost==1.2.2
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
pyyaml==6.0
matplotlib==3.8.0
shap==0.45.0
shap
PyYAML
